Make SVG map tester run against all variants.
The generated maps for the Classical map all look wrong because it contains
a translation.  It should be fairly easy to remove this translation and the
map will start behaving as expected. Also non-supply center provinces are
shown in black on some maps, but I'm not sure that's an issue.
